requirements.txt
Appium-Python-Client==5.2.4
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
pycparser==2.23
PySocks==1.7.1
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
